+++ GameParticles/Particle.cpp
@@ -30,7 +30,7 @@ void Particle::CopyDataOnly( Particle *p )
 	this->life     = p->life;
 }
 
-void Particle::Update(const double& time_elapsed)
+void Particle::Update(const float& time_elapsed)
 {
 	// Rotate the matrices
 	Vect4D tmp_Row0;
@@ -56,7 +56,7 @@ void Particle::Update(const double& time_elapsed)
 	tmp.set(Matrix::MATRIX_ROW_2, &this->diff_Row2);
 	tmp.set(Matrix::MATRIX_ROW_3, &this->diff_Row3);
 
-	double MatrixScale = tmp.Determinant();
+	float MatrixScale = tmp.Determinant();
 
 	// serious math below - magic secret sauce
 	life += time_elapsed;
@@ -65,14 +65,14 @@ void Particle::Update(const double& time_elapsed)
 	Vect4D v(3,4,0);
 	position.Cross( z_axis, v);
 	v.norm(v);
-	position = position + v * 0.05 * life;
+	position = position + v * 0.05f * life;
 
-	if( MatrixScale > 1.0 )
+	if( MatrixScale > 1.0f )
 	{
-		MatrixScale = 1.0/MatrixScale;
+		MatrixScale = 1.0f/MatrixScale;
 	};
 
-	rotation = rotation + MatrixScale + rotation_velocity * time_elapsed *2.01;
+	rotation = rotation + MatrixScale + rotation_velocity * time_elapsed *2.01f;
 }
